+++ helpers.test.js
@@ -124,11 +124,11 @@ describe("Helpers", () => {
       const lastMessageSystem = { role: "system" };
       const resultUser = isBreakingMessageChain(
         capabilityMatch,
-        lastMessageUser,
+        lastMessageUser
       );
       const resultSystem = isBreakingMessageChain(
         capabilityMatch,
-        lastMessageSystem,
+        lastMessageSystem
       );
       expect(resultUser).toBe(false);
       expect(resultSystem).toBe(false);
@@ -152,7 +152,7 @@ describe("Helpers", () => {
         expect(messages[0].role).toEqual("system");
         // Adjusted the regex to match the received string format including AM/PM
         expect(messages[0].content).toMatch(
-          /Today is \d{1,2}\/\d{1,2}\/\d{4} at \d{1,2}:\d{2}:\d{2} [AP]M/,
+          /Today is \d{1,2}\/\d{1,2}\/\d{4} at \d{1,2}:\d{2}:\d{2} [AP]M/
         );
       });
     });
